Corrected some problems with Cij augmentation and discovered that it needs scaling to work well

+++ src/mumps/sparse_simple_sumproject.cpp
@@ -217,7 +217,7 @@ void abcd::spSimpleProject(std::vector<int> mycols, std::vector<int> &vrows,
 
                     if(ci >= col && val != 0){
                         if ( ci == col)
-                            val += 0.5;
+                          val += (double)0.5;
 
                         vvals.push_back(val);
                         vrows.push_back(ci + 1);
